+++ Source/Bake/Services/Tools/DotNetArguments/DotNetPublishArgument.cs
@@ -20,6 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+using Bake.Core;
 using Bake.ValueObjects;
 
 namespace Bake.Services.Tools.DotNetArguments
@@ -32,6 +33,8 @@ namespace Bake.Services.Tools.DotNetArguments
         public string Configuration { get; }
         public Platform Platform { get; }
         public string Output { get; }
+        public SemVer Version { get; }
+        public IReadOnlyDictionary<string, string> Properties { get; }
 
         public DotNetPublishArgument(
             string filePath,
@@ -40,7 +43,9 @@ namespace Bake.Services.Tools.DotNetArguments
             bool build,
             string configuration,
             Platform platform,
-            string output)
+            string output,
+            SemVer version,
+            IReadOnlyDictionary<string, string> properties)
             : base(filePath)
         {
             PublishSingleFile = publishSingleFile;
@@ -49,6 +54,8 @@ namespace Bake.Services.Tools.DotNetArguments
             Configuration = configuration;
             Platform = platform;
             Output = output;
+            Version = version;
+            Properties = properties;
         }
     }
 }